Fix some of the warnings in macosx SITL build
Mostly promotion from float to double, since OSX is a 64bit build.

+++ src/main/common/maths.h
@@ -188,6 +188,6 @@ float acos_approx(float x);
 void arraySubInt32(int32_t *dest, int32_t *array1, int32_t *array2, int count);
 
 float bellCurve(const float x, const float curveWidth);
-float fast_fsqrtf(const double value);
+float fast_fsqrtf(const float value);
 float calc_length_pythagorean_2D(const float firstElement, const float secondElement);
 float calc_length_pythagorean_3D(const float firstElement, const float secondElement, const float thirdElement);
