+++ src/main/telemetry/sim.c
@@ -375,7 +375,7 @@ static void sendSMS(void)
         amps / 10, amps % 10,
         getAltitudeMeters(),
         groundSpeed, avgSpeed / 10, avgSpeed % 10,
-        GPS_distanceToHome, getTotalTravelDistance() / 100,
+        (unsigned long)GPS_distanceToHome, getTotalTravelDistance() / 100ul,
         DECIDEGREES_TO_DEGREES(attitude.values.yaw),
         gpsSol.numSat, gpsFixIndicators[gpsSol.fixType],
         simRssi,
